requirements.txt
Flask[async]==2.0.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
Flask-Mail==0.9.1
Flask-Caching==1.10.1
Flask-Login==0.5.0
Flask-Admin==1.6.0
email-validator==1.1.3
itsdangerous==2.0.1
pytest==7.0.1
python-dotenv
pathlib2
qrcode